+++ src/lib.rs
@@ -1,11 +1,7 @@
 // Copyright (c) 2023 xplm-sys developers
-// Licensed under the Apache License, Version 2.0
-// <LICENSE-APACHE or
-// http://www.apache.org/licenses/LICENSE-2.0> or the MIT
-// license <LICENSE-MIT or http://opensource.org/licenses/MIT>,
-// at your option. All files in the project carrying such
-// notice may not be copied, modified, or distributed except
-// according to those terms.
+// SPDX-FileCopyrightText: 2024 Julia DeMille <me@jdemille.com
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
 
 // Allow C-like conventions
 #![allow(non_upper_case_globals, non_camel_case_types, non_snake_case)]